request #14471: Update to guzzle/psr7 1.6.1
A change is made to the FileUploadController so the implementation
of the TusServer can continue to work. In guzzle/psr7#247 [0] the default
body of the ServerRequest has been switched to a CachingStream which is
nice until you need to use StreamInterface::detach() because only already
cached data will be available. Forcing data to be put in cache is not interesting
here because we want to save as much as data as possible from the client even
if we only got a part of it (resuming the upload will be dealt with the tus protocol).
To avoid that, we now recreate a StreamInterface directly from the php://input resource
when uploading a file with tus.

[0] https://github.com/guzzle/psr7/pull/247

Change-Id: I167bac3958ffdec37d8b4cf8860d70d73befe3a1