Do not add AddToTopBacklog post action if tracker is not top backlog tracker
This is part of story #14096 have a post action to add "Scrum Top Backlog" flag

If the tracker is not a top backlog tracker, then the UI to add the
AddToTopBacklog post action in the workflow must not be displayed.

In addition, the REST backend must block the call if a AddToTopBacklog
post action is provided if tracker is not top backlog tracker.

Change-Id: I22694f1722b224e064611032f4cd353a3d84e66d