request #14484: FieldVisitor is incomplete
When Tracker_FormElement_Field_Priority was created, the existing
Tracker_FormElement_FieldVisitor had not been updated. Therefore there
can be cases where the Priority field can call visitPriority() method on
a visitor that does not implement it.

This has been kept under the radar because Priority field is a readonly
field (no XML import) and is not provided in the TQL expert mode (both
cross tracker and tracker report). However if you entered manually the
name of the priority field in the TQL interface you could get a fatal
error.

For example: have a priority field named "rank". Go to TQL and search
for "rank > 1".
=> You should get an error message saying that "rank" is not supported,
instead of a fatal error.

Change-Id: Icd0a1eebcb95724a26c0a113256efdc5c463266f