request #14488 Select2 dropdown is under the artifact modal
How to test:

In kanban (or another app using the artifact modal), set up a tracker
with:
- an open list
- a date field

When you edit an artifact via the artifact modal, when you type in the
open list field, the select2 dropdown should be above the modal and not
under it.

No other functional change expected.

Change-Id: I75862fdc5fd4b8de4476099871d11b4037d2d917