request #14489: Help text is displayed as placeholder in new project creation
Placeholders should not be a long paragraph. It should easily give clues
of the purpose of the field. Any instructions or help text should be
displayed alongside the field. Else it can produce weird displays
(display of html tags, unreadable placeholder if long description, …).

Description of project fields should be treated as description and must
be placed under the text field, as tlp-text-info.

In order to not clutter much the creation page, now the description of
project fields is not anymore mandatory. If empty, then no tlp-text-info
will be displayed.

Change-Id: I2b44bf3e3bdbf3a8aaeca00e26d90734a01153a3