Agile ALM template should be taskboard friendly
Create a new project with "Agile ALM" template.
Go to the project » agiledashboard.
Create a milestone with a user story.
Go to the taskboard of the story.
=> you should be able to add in place tasks for the story.

Bonus: fix the post project creation illustration (when project needs
siteadmin approval) that had wrong classnames (due to an issue in the
illustration helper in tlp)

Part of request #14490: Default Scrum Template is not compatible with Taskboard

Change-Id: I59c04263477f41b6be66444ca7620ab0872fa83b