request #14524: Remove unused localized label in StaticText
In StaticText, some code is useless because it comes from the legacy TV3
system. There was no static text in TV3 therefore there is no chance to
retrieve them in TV5 after a migration.  This is not the case for
fieldset though, therefore we should probably keep
isLabelMustBeLocalized() (however there is no call to
fieldset->getDescriptionText, it can be removed)

Change-Id: Ia1c102be686063d0896034759193960a5b46a003