request #14529: Do not depend on concrete logger implementations
To respect the dependency inversion principle so we easily swap
implementations of those loggers in the future, the code now depends
on the abstraction everywhere it was possible to do the switch.
The loggers in the tracker workflow area are kept as is because they
define methods outside of the PSR-3 LoggerInterface, the implementation
cannot be swapped for something else.

Change-Id: I56224bd3736d7be55dd6ee1a616c6e7a915a03db