request #14534: Refuse CSV import file that are not in UTF8
When you import a CSV file in tracker, the validation of the file might
fail if the encoding of the file is not UTF-8.

For example, create a static list with a value containing accentuated
characters like "données applicables".  If your CSV file is in
ISO-8859-1 and contains this value, then you will get an empty error
message. (Purifier mangled the error message due to the presence of non
UTF-8 characters).

In order to mitigate that, I suggest that we now refuse any file that
are not UTF-8 (us-ascii are still accepted though).

Change-Id: I34cf5b4f3056f93d4bad2b22ac93cfd5223a9b72