inject trackers into DOM
part of story #14291 Beautiful tracker creation

Why this change:
In nexts steps we'll need to display a dropdown list with trackers group by project.

This patch introduce the backend mechanism of retrievement of this lists.

How to test
Make sure that the vue app is not built
check the value of data-project-templates of div tracker-creation-app

You must have a list of trackers grouped by project

Change-Id: Ie3cafbd5a73ce930727800035da081c1529c5cc7