Version of the label plugin RPM package is incorrect
The version for this plugin is the one of the dynamic credentials
plugin. Due to this situation, updates cannot be achieved because
the dynamic credentials plugin ash a lower version than the label
plugin so the new label package is ignored.

request #14520: Merge the specfile of the Enterprise plugins into
the main specfiles

Change-Id: Idd6a9051ad6f1672be49cb830d207f144b5ae9b7