Add a new pane in Git administration when hudson_git is available
This is part of story #14019 automate jenkins job creation & run

You have to set a new feature flag to 1 to enable this feature:
$git_allow_jenkins_plugin_hook_configuration = 1;

When this is enabled and the hudson_git plugin is installed and
activated, then a new pane named "Jenkins" must be displayed in the Git
administration page of a project.

The pane does nothing for now. Its content will be added in a dedicated
commit.

Change-Id: Idc63c9e724545be540819fd334e9cdc4cc476f67