Refactoring: Extract project retriever
Part of story #14541 have OAuth2 project admin

Project admin has more and more controllers. They should all check that
the given project is not null and not in error.
This contribution introduces a dedicated object for that purpose.

The GetProjectTrait usage has also been replaced. The Trait is fine
except that the IDE does not warn us when we forget injecting
ProjectManager in the constructor. I decided to go with the regular
Class.

No functional change expected in project admin or FRS licenses admin.

Change-Id: I03587af6f3fa3c1d7e1ecfbe44dc0315965425c0