Move initialization of Tuleap DB to tuleap-cfg commands
First and foremost the goal is to be able to understand what the code
does.

This commit introduce 2 tuleap-cfg commands:
- One for database initialization. It assume a working DB with enough
  privileges to create DB and users with all privileges grant.
- The second one assume that the DB and app user exists and load
  everything in the DB.

The need is to be able to run the second only in setup where the DB
is managed by an external service (like Cloud managed DBs).

Part of: story #14683 install a basic image

Change-Id: Iaa127beaae8cf3434b2ac70ed25c1f8999b52f3f