Convert TrackerXmlImportTest into phpunit
part of story #14291 Beautiful tracker creation

Why this change?
I plan to inject TrackerXmlImportFeedbackCollector as dependency of class (instead of making new in code).
The way the simpletests were written was no help to do easilly such thing, and we are migrating all tests to phpunit.

I choosed to do the test migration before updaing any code.

This legacy code were always testing from `import` method
who was painfull to understand and maintain.

The better way I found to don't loose my brain and don't remove all tests was to split things we are testing in protected method and do the code and the assertion only on this part.

Change-Id: I515dcb4d60f80e681add0fd677e7261839060fb9