Protect a test resource with an OAuth2 access token
In order to progressively implements the whole OAuth2 flow,
a test resource is created at '/plugins/oauth2_server/testendpoint'.
Once the implementation is done, this resource will be removed.

You need to provide a valid OAuth2 access token as a Bearer token [0]
to access it. For now there is only one valid hardcoded access token:
'tlp-oauth2-at1-1.6161616161616161616161616161616161616161616161616161616161616161'.

The access token is not yet associated with the notion of scopes.

The middleware does not yet properly sets the WWW-Authenticate header
when the OAuth2 access token can not be validated.

To test it, you can use curl as follow:
    curl \
        -H 'Authorization: Bearer tlp-oauth2-at1-1.6161616161616161616161616161616161616161616161616161616161616161' \
	https://tuleap.example.com/plugins/oauth2_server/testendpoint

This is part of story #14542: have OAuth2 flow

[0] RFC6750: https://tools.ietf.org/html/rfc6750

Change-Id: If21bbd0edd36b8d535722e3283223844d108ec54