Add a bunch of builder to ease test of controllers
Testing controllers can be a pita because of injections and
dependencies on hard to build objects.

This introduce a bunch of builders to use real objects instead
of mocks. By doing so, we avoid introducing logic in tests, tests
are more simple to read and focuses on the intend rather than
duplicating target code logic.

Part of story #14552 User preferences goes Burning Parrot

Change-Id: I0f02d1f8370410e9bc866e3fb76aac731903a1c6