Introduce tlp-tabs-vertical
We can now have vertical tabs in TLP. Go to TLP documentation »
structure & navigation » tabs to see an example of usage.

Bonus: every links in the examples didn't work anymore due to
sanitization introduced earlier. href="javascript:;" was used to provide
valid link, but sanitize() consider this to be harmful, leading to a <a>
tag without any href attribute. This is problematic because some
navigators don't consider it anymore as a hyperlink tag, and behavior
can change (like no pointer cursor on hover).
Now example links are using href="#", but to not force the browser to
follow the dummy anchor (typically scroll back to the top of the
document), we disable the click on the link.

Part of story #14552 User preferences goes Burning Parrot

Change-Id: I9311e4c1b39ae827e26e3a72a78e7d507a1b471a