OAuth2 access token are associated with scopes
Each resource accessible with an OAuth2 access token has
a specific scope linked to it.
For now the test endpoint introduced in 976a0efcc05dd92504875e66c14b6580a67c32c4
has been reused and is retricted to the 'demo' scope.

To test you need to create an access token with the required scope:

    mysql > INSERT INTO oauth2_access_token (user_id, verifier) VALUES(<user_id>, SHA2('aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa', 256));
    mysql > INSERT INTO oauth2_access_token_scope (access_token_id, scope_key) VALUES(<token_autogenerated_id>, 'demo');

    shell > curl \
        -H 'Authorization: Bearer tlp-oauth2-at1-<token_autogenerated_id>.6161616161616161616161616161616161616161616161616161616161616161' \
	https://tuleap.example.com/plugins/oauth2_server/testendpoint

The definition of the scopes will be reused by refresh tokens and grants
for the authorization code flow.

To test with an invalid required scope, you edit the entry you
added in the oauth2_access_token_scope table or remote it.

This is part of story #14542: have OAuth2 flow

Change-Id: Ic09e4f16dcab544740c8f51cf8f5fdb1c86ee173