Fix unit test dependency
With usage of includeFooterJavascriptFile the target file must
exist on filesystem when the call is made. However it should
be possible to run unit tests without having built assets so
tests were failling.

This commit introduce addJavascriptAsset (mirror of addCssAsset)
so resolution of the file can be defered to the very last moment
hence unit tests passes again.

Part of story #14552 User preferences goes Burning Parrot

Change-Id: Iecc64a34d9fb7d236489e248c9205021380e5981