Add legend points and time
This is part of story #13301 Begin BurnUp

Please, for the test, you need to set '$project_milestones_activate_burnup = 1;' in your local.inc.

Now, there are points and time on scale of the burnup.
The time scale must be the same as on the burndown (of the widget).
The points scale should be the same as on AD.

How to test:
	- Have project milestones plugin
	- Have some releases with burnup

Expected results:
	- Burnup has labels on scale.
	- Same time scale of burndown
	- Same points scale of burnup in AD.

Change-Id: Ie2cd5c5ffb0c3551936850dd53cbb4c0238b04fc