Base of authorization endpoint form
Part of story #14570 Authorization grant confirmation page

How to test:
- Go to a project's administration
- Create an OAuth2 app
- query its ID in the database
- go to
https://tuleap.example.com/oauth2_server/authorize?client_id=tlp-client-id-<app_id>

A form should be rendered with the matchin App's details. Scopes
(permissions) are all fake for now and will be filled in a later
contribution. Buttons are disabled.

If you go to this URL as anonymous, you should be redirected to the
login page.

Invalid client_id should be forbidden.

Change-Id: I8e8724a99dfbe4221dbecf9a815d50f6b443b897