Create the base architecture to create an OAuth2 access token from an authorization code
For now, an access token is issued for the 'admin' user with a 'demo' scope
to anyone doing a POST request on the route. This token can be used against
the test resource route.

No verification is done before issuing the OAuth2 access token. All the
required verification will be progressively added in future contribution.

You can test as follow:

    shell> curl -X POST https://tuleap.example.com/plugins/oauth2_server/access_token
    shell> curl \
        -H 'Authorization: Bearer <token_retrieved_from_the_previous_response>' \
        https://tuleap.example.com/plugins/oauth2_server/testendpoint

This is part of story #14542: have OAuth2 flow

Change-Id: I71430653dd249e8c7099915dc2a902b5f4834800