request #14573 Cypress tests should not rely on past test execution
When we introduce cypress we had a batter test patten:
=> permissions tests were relying on agiledashboard tests
(Ad was created by UI and permissions on ad were checked later).

With this patch the AD is instantiated by UML.

How to test:
make test_cypress
you should not have errors in /tests_results_e2e_full

Please note:
yes I remove the start scrum test, it will be added again later with gerrit #16503
we should not be afraid to break XML import during the time we won't have the removed test on master.
Agiledashboard XML creation is still tested in cypress when we do a start scrum,
the only thing that this removed test, might not catch is if the AD empty state is broken.

Change-Id: I497bb8dba6ebbda97c49865740e2c238c5711897