Clear all OAuth2 access tokens and associated scopes when the OAuth2 server plugin is uninstalled
Those tables are managed by the Tuleap core to ease the management
of the authentication system. However by themselves they are not
useful, the only way to create an OAuth2 access token is through the
OAuth2 server plugin. As such should be cleared when the plugin is
uninstalled.

This is part of story #14542: have OAuth2 flow

Change-Id: I5cab77c3ce301586eb0e97569c923f195849aecb