Do not import in a subfolder of root item
When we export a project, the first item in the <docman> is the root
folder.

When we import, this root folder does not need to be imported, because
it should already exist in the project[0]. We should not have a root folder
with the imported root folder right under it.

[0] If for some obscur reason the project does not have a root folder,
then it means that something goes wrong during project creation. In that
case the import is not done. We could try to (re)create the root folder
however the chance to end up in this situation is very small. Therefore
I let it as an exercice for our future us if the issue become apparent.

Part of request #14507: XML import/export : docman

Change-Id: I0cc8efa0603dc8896c1f50d4d96e6636e6d5116a