The authorization page should not be cached or accessed without TLS
This page is part of the authorization server the same constraints
than for the endpoint creating the OAuth2 access token must be applied.

In order to reuse the same code between the two pages, the controller
of the authorize page can now be used with PSR-15 middlewares. The integration
is however not really pretty due to the way the themes are currently rendered.

This is part of story #14542: have OAuth2 flow

Change-Id: Idcf40cca32330dbd26ff959f05950abdeaec4892