request #14588: Remove versions of plugins and themes that are locked on the main Tuleap version
As of today all the plugins and themes maintained in the main Tuleap repository
have each an independent version. This serves no purposes as all these packages
are directly dependent of the main Tuleap version.

Removing them reduce the boilerplate needed when creating a new plugin and
it brings us one step closer to simplify the integration of contributions.

The VERSION files are kept with this contribution even if they are no more
useful to avoid permanent conflicts with the current HEAD of the master
branch. They will be removed in a second contribution.

Change-Id: I0515bdc247d41cec8376fa674ef449808831110a