Update password within the new Account page
It's only mainly the backend part, the helpers on front end
side will come in a dedicated patch.

The old page is removed and the related features too (like
the change of password when expiration is reached).

The old change_pw.php was accessible by anonymous by accident
in order to allow change of password after expiration. It's
now properly managed with an explicit whitelisting of the
2 allowed URLs.

Part of story #14552: User preferences goes Burning Parrot

Change-Id: I735f16d0bc861c172857ceb8d4cdd18ec036c0a2