Each OAuth2 app is assigned a secret when it is created
The secret is generated automatically and is stored in the database
in a non reversible way and is shown to the user creating the app
once.
For now only one secret can be created by app.

The SQL queries retrieving the app information to not expose
the hash of the secret when it is not absolutely required.

Part of story #14541: have OAuth2 project admin

Change-Id: Ie0cab22c24eb6c75c192a879dcb954e970aaf12b