Improve XML file parsing
Now when the user selects a file that is not a XML file,
The app does not attempt to parse it. For the record, Edge Legacy
does not support the "accept" attribute on inputs, so when the OS
file manager opens, it does not filter on XML files.

Also a spinner was missing on the [Next ->] button while data
was being extracted from the xml file. When it takes a longer time,
the button is clickable although the data have not been extracted yet.

story #14291 Beautiful tracker creation

How to test:
- Try to upload a big, non-xml file
--> The error is displayed immédiately (no delay)

- Slow down the perfs of your browser, and select a big xml file
--> The [Next ->] button is disabled
--> A spinner replaces fa-long-arrow-right

Change-Id: Ie1fd2d9cb1f522549e29fc45d7b85a41706c6840