Introduce password update interface
In order to reduce the amount of code to review, this first patch
only introduce the new UI with all fields disabled + the logic of
display with a new Event.

The un translated string to warn people about work in progress was
made on purpose because will be renamed in the next patch.

Part of story #14552: User preferences goes Burning Parrot

Change-Id: I9f7a65ad5963903b9b930596c04f8d7e9a64b635