Rely on PSR-17 interface instead of concrete implementations
The URI built to redirect the resource owner to the client
after its approval/denial of the requested authorization was
built using directly the library for our PSR-7 implementation.

Moving to the PSR-17 interface allow us to respect the dependency
inversion principle (and it simplifies the code a bit).

The link on the "Deny" button of the authorization page should stay
the same. No functionnal change is expected.

To test:
 1. Go to the admininistration of one project and create an OAuth2
with "https://example.com" as the redirect endpoint
 2. Go to the authorization page:
https://tuleap.example.com/oauth2_server/authorize?client_id=<client_id>&state=xyz&scope=demo%20read:project&response_type=code&redirect_uri=https%3A%2F%2Fexample.com

Part of story #14570: Authorization grant confirmation page

Change-Id: Ib6cc1afc5a047dc12cc8c4288098f94cf351619c