Attempt to wipe sensitive strings from memory once they are not more needed
ConcealedString tries to clean up the wrapped secret. This is a best practice
to try to limit as much as possible the exposure of sensitive information.

For data coming from or being sent in a HTTP request it is not really possible
to scrub everything from memory. However this change allow us is interesting
for the Tuleap internal crypto API where everything is manipulated locally.

No functionnal change is expected.

Part of request #14602: Harden handling of sensitive strings

Change-Id: I1869700c4513ff1d5c5d1ff177bcecb134f586cb