Remove user_legal
An hidden feature was introduced in 2011, but has never been really
used, documented, nor promoted. This "feature" allows an administrator
to display a bunch of text on the register and account pages.
Accordingly to the name of the involved variables/files, it relates to
some legal notice to display to end users.

Since there is no usage of this we should remove it. If the need emerges
to display legal notice to end users, then it should be done properly,
not by displaying unsanitized random stuff in the middle of the page.

Part of story #14552: User preferences goes Burning Parrot

Change-Id: I102d5a5cf1fa45967127c4f2562c79c7eab6c22e