Be consistent in the endpoint URLs of the authorization server
This contribution changes the URLs of the authorize and token
endpoint. Both are supposed to be used in the same flow, it is
nicer for someone implementing an OAuth2 client if both URLs
have the same format.

The authorize endpoint is now at /oauth2/authorize.
The token endpoint is now at /oauth2/token.

Part of story #14542: have OAuth2 flow

Change-Id: I026ce4d45df0d146960e9cdb8f77ccb9391cfd18