request #14612: unterminated entity reference during tracker export
If the name of a report contains ampersand, then the export of the
project should not display a warning.

To reproduce the issue you can use the following string "R&D" in report
name, report description, renderer name, …

We should probably add a psalm check (or other) to prevent future unsafe
usage of SimpleXMLElement->addChild(). This should come in a dedicated
changeset however.

Change-Id: Ie80cb98ca5a3075c7eeaf2869971e810b4959216