Fix fatal error on account » security
If you are a user that never fails to login, then you should not get a fatal
error on account » security page.
This was due to DateHelper::formatForLanguage() which returns null in
that case (and mustache fails).

Bonus:
- rename the pane title to "Access history",
- group "successful login" properties,
- remove useless nb last login failure.

The nb last login failure is useless:
- This number increases each time the user fails to login.
- It is set to 0 as soon as the user logins successfully.
- => it is always 0 => it is useless for the end user.

Part of story #14552: User preferences goes Burning Parrot

Change-Id: I0764bf581922b36b864fbc0b141f42c8d0ce16dc