Move code related to the authorization code when granting an access token into a subnamespace
This contribution does not bring any new functional changes.
It only reorganize a bit the content of the Tuleap\OAuth2Server\Grant
namespace before adding more code to it.
In a close future access token will also be granted with a refresh token.
It's easier to the changes now when there is not that much code than later.

This is part of story #14542: have OAuth2 flow

Change-Id: Id09e64da8e4024be342f12ca230d28a979a9def4