Defer the display of password robustness
No need to alarm the users by saying that the password is weak whereas
they only display the security section.

Now the password robustness is displayed as soon as the user focus the
new password field. We should not wait for users to type something
because we should provide guidance about expected password format as
soon as possible. Focus seems a good tradeoff.

Part of story #14552: User preferences goes Burning Parrot

Change-Id: If4b02af64735e9b60902343d1f8c4f3c0c32355e