request #14637: Remove usage of the Jenkins clover plugin the test pipelines
The plugin is a source of issue [0] which is unlikely to get some
attention soon [1]. It also sometimes crash when analyzing the reports
which completely fail the pipeline even if all the tests have been passed
successfully.

The Javascript now uses the modern approach to do code coverage in
Jenkins [2].
For the PHP code coverage the situation is less pretty. PHPUnit cannot
output a report in a format understood by the Jenkins plugin which means
we loose the high level overview. The HTML report is still browsable
from the archive and it looks like there is currently some traction to
add a format that would be supported by the Jenkins plugin [3] (we can
also consider writting a small tool to do the format conversion).

[0] https://github.com/jenkinsci/clover-plugin/pull/21
[1] https://github.com/jenkinsci/clover-plugin/pull/21#issuecomment-583120076
[2] https://jenkins.io/blog/2018/08/17/code-coverage-api-plugin-1/
[3] https://github.com/sebastianbergmann/php-code-coverage/pull/734

Change-Id: Ic4915047616a7023515939dd305a23f12f2300cc