request #14636: Scrum templates should have comfortable edit zone
When you create an AgileALM project, or a Scrum project (or you click on Start scrum button),
Then the epic and user story trackers should not have textearea on the same line.

When you write acceptance criteria, you want a comfortable edit zone, not a too small editor.

PainPoints++: we should not have two different xml files for the same
tracker.

Change-Id: Id1548e4504e566d5ec7418b517f587b38ee83af7