Introduce xsltproc to generate XML templates
Until now there was 3 different XML templates for bug tracker, 3
different XML templates for user story, and 3 different XML templates
for task tracker. This means that every time we enhance a tracker, the
modification should be done everywhere so that user have consistent
results.

The proposal here is to use XSLT to generate different templates.

You need to run the following command to generate templates
$ make generate-templates
or, if you do not have xsltproc:
$ make generate-templates-docker

The generated templates does not bring any functional changes.

The templates are still in the sources in order to not break the CI. Here
is the roadmap:
- Generate templates via XSLT (this contribution)
- Update CI so that it generates templates (see pull request[0])
- Remove templates from the sources (next contribution[1])

Note: as of today there are still individual templates that are duplicated:
plugins/tracker/www/resources/templates/Tracker_{Bugs,Tasks,UserStories}.xml
A dedicated contribution will deal with them.

Part of request #14654: Do not repeat trackers XML templates

[0] https://tuleap.net/plugins/git/?action=pull-requests&repo_id=227&group_id=101#/pull-requests/103/overview
[1] https://gerrit.tuleap.net/c/tuleap/+/18163

Change-Id: Iaa565ca131af938b0e28e25647459574e4cf95f6