request #14642: Feedbacks displayed on homepage are ugly
If a feedback is displayed on the homepage, then it is displayed on a
black background, and is barely readable.

This can be reproduced on master by going to account page and asking for
an email change. Then follow the link in the confirmation email but
change the confirm hash so that the request is invalid.

Change-Id: I7a5eb71ca9dab896ff04f671ef9b1c8a38428f4f