request #14655: Parallelize the builds of the assets
Lerna is used to achieve that as a replacement of gulp.

The concurrency is limited to <number_cpus> - 2 to keep some breathing
room when building the assets.

To manage the dependency between internal packages, the dependency
are now defined in the package.json with a local path.
However when doing that npm adds the whole tree of the local dependency
into the lockfile which makes things hard and fastidious to maintain.
When Lerna is used to completely manage the dependencies the lockfiles
does have the mention of the local dependency as one would expect.
However in the situation in which Tuleap is with a composite mono-repo
where some parts might not present it is not possible the use Lerna with
all its capabilities. Hoisting becomes impossible and overall managing
dependencies at the global level is dangerous because we are never sure
if we have the complete snapshot of the monorepo or not. To avoid that
we keep managing each package individually and for packages with a local
dependency we rewrite the lockfile to not have duplicate trees. As a
result usual npm commands still work (with the exception of 'npm ci'
since the lockfile appears incomplete) and contributors does not have to
learn another tool. Since the local dependency are defined in the
package.json Lerna is capable to run the build scripts in the order
imposed by the dependencies.

Change-Id: I62b625e4f7f7a150f6afc4325ab16b83294dbc9f