Transform remaining Git/Gitolite driver SimpleTest tests in PHPUnit integration tests
Those tests have far too much side effects (I/O accesses everywhere) to
be considered as unit tests and they need access to a version of Git
similar to what's available on production environnements.

With this contribution, the Git plugin is now free of SimpleTest.

This is part of request #14150: Get rid of SimpleTest

Change-Id: Iee87555ac10b9fd97d65bd500c78fa626bcb0ea4