request #14669: Sidebar in siteadmin should not be collapsable
The user experience of the collapsed sidebar in site administration is
not really nice :

- We need to expand back the sidebar to navigate to other sections of
site administration
- There is no visual clue on how to expand the sidebar
- Users need to click on a 10px wide zone

Since we don't have much time to spend on the improvement of this, and
this prevents people to work, we suggest to remove the capacity to
collapse the sidebar.

The sidebar is still collapsable in projet workspace though.

Change-Id: I8bc0ed68cfdb494b9108cc7519fcfac715f5a0b5